c backend: handle int division

+++ tests/machine/x/c/math_ops.c
@@ -3,7 +3,7 @@
 
 int main() {
   printf("%d\n", 6 * 7);
-  printf("%d\n", 7 / 2);
+  printf("%g\n", ((double)7) / ((double)2));
   printf("%d\n", 7 % 2);
   return 0;
 }
